Refactoring card reading strategies into separate files
