+++ munchkin/core/strategies/bottom_to_top.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+
+def bottom_to_top(card):
+    """ If the password from the card is read from bottom to up """
+    data = card.getT().getA1().flatten()
+    results = data.tolist()
+    results.reverse()
+    return results
