+++ munchkin/core/strategies/left_to_right.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+
+def left_to_right(card):
+    """ If the password from the card is read from left to right """
+    data = card.getA1().flatten()
+    results = data.tolist()
+    return results
