+++ munchkin/core/strategies/right_to_left.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+
+def right_to_left(card):
+    """ If the password from the card is read from right to left """
+    data = card.getA1().flatten()
+    results = data.tolist()
+    results.reverse()
+    return results
