+++ munchkin/core/strategies/zig_zag_reverse.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+
+import itertools
+
+def zig_zag_reverse(card):
+    """ If the password from the card is read in reverse zig-zag directions """
+    rows = card.getA().tolist()
+    rows.reverse()
+    for i in range(len(rows)):
+        # Reverse order on every second line
+        if i % 2 == 0:
+            rows[i].reverse()
+    # Flatten list
+    results = list(itertools.chain.from_iterable(rows))
+    return results
